Misc refactors
